fix: only update active MemberProfiles with affiliations
- use requests.Session context manager instead of caching on the Command
  instance
- add with_affiliations() queryset method to MemberProfile and other
  minor hygiene
